app.js added, index.js is more simple now
